refactor: extract HTTPAPITypeDefinition to a separate file
